Refactor controller components and update imports for improved organization
- Reorganized imports in App.tsx to reflect new directory structure for controller components, enhancing clarity and maintainability.
- Removed unused components: ControllerButtons, ControllerGameHub, ControllerLibraryPage, ControllerStreamLoading, and ControllerInStreamShell, streamlining the codebase and reducing complexity.

These changes contribute to a cleaner architecture and improved component management within the application.
