+++ opennow-stable/src/renderer/src/App.tsx
@@ -58,11 +58,11 @@ import { LoginScreen } from "./components/LoginScreen";
 import { Navbar } from "./components/Navbar";
 import { HomePage } from "./components/HomePage";
 import { LibraryPage } from "./components/LibraryPage";
-import { ControllerLibraryPage } from "./components/ControllerLibraryPage";
-import { ControllerInStreamShell } from "./components/controllerInStream/ControllerInStreamShell";
+import { ControllerLibraryPage } from "./components/controllerMode/ControllerLibraryPage";
+import { ControllerInStreamShell } from "./components/controllerMode/controllerInStream/ControllerInStreamShell";
 import { SettingsPage } from "./components/SettingsPage";
 import { StreamLoading } from "./components/StreamLoading";
-import { ControllerStreamLoading } from "./components/ControllerStreamLoading";
+import { ControllerStreamLoading } from "./components/controllerMode/ControllerStreamLoading";
 import type { QueueAdPlaybackEvent, QueueAdPreviewHandle } from "./components/QueueAdPreview";
 import { StreamView } from "./components/StreamView";
 import { QueueServerSelectModal } from "./components/QueueServerSelectModal";
