+++ opennow-stable/src/renderer/src/components/controllerMode/controllerLibrary/useControllerLibraryEvents.ts
@@ -1,0 +1,309 @@
+import { useMemo } from "react";
+import { createActionHandlers } from "./events/actionHandlers";
+import { createNavigationHandlers } from "./events/navigationHandlers";
+import type {
+  ControllerLibraryEventContext,
+  ControllerLibraryEventHandlers,
+} from "./events/types";
+export type { ControllerLibraryEventHandlers } from "./events/types";
+
+export function useControllerLibraryEvents(
+  ctx: ControllerLibraryEventContext,
+): ControllerLibraryEventHandlers {
+  const {
+    isLoading,
+    TOP_CATEGORIES,
+    categorizedGames,
+    selectedIndex,
+    selectedGame,
+    selectedGameId,
+    selectedVariantId,
+    onPlayGame,
+    onSelectGameVariant,
+    onToggleFavoriteGame,
+    playUiSound,
+    throttledOnSelectGame,
+    topCategory,
+    selectedSettingIndex,
+    selectedMediaIndex,
+    selectedGameSubcategoryIndex,
+    displayItems,
+    mediaAssetItems,
+    mediaSubcategory,
+    gameSubcategory,
+    settings,
+    settingsSubcategory,
+    lastRootSettingIndex,
+    lastRootMediaIndex,
+    lastRootGameIndex,
+    lastSystemMenuIndex,
+    lastThemeRootIndex,
+    onSettingChange,
+    resolutionOptions,
+    fpsOptions,
+    codecOptions,
+    aspectRatioOptions,
+    currentStreamingGame,
+    onResumeGame,
+    onResumeCloudSession,
+    onCloseGame,
+    onExitControllerMode,
+    onExitApp,
+    editingBandwidth,
+    editingThemeChannel,
+    gamesShelfBrowseActive,
+    mediaShelfBrowseActive,
+    topLevelRowBehaviorActive,
+    topLevelShelfIndex,
+    canEnterDetailRow,
+    canEnterTopRow,
+    ps5Row,
+    detailRailIndex,
+    detailRailItems,
+    optionsOpen,
+    optionsFocusIndex,
+    optionsEntries,
+    gamesRootPlane,
+    spotlightIndex,
+    spotlightEntries,
+    gamesDualShelf,
+    favoriteGameIdSet,
+    microphoneDevices,
+    gamesHubOpen,
+    gamesHubFocusIndex,
+    gamesHubTiles,
+    inStreamMenu,
+    endSessionConfirm,
+    editingStreamVolume,
+    editingStreamMicLevel,
+    streamMenuMicLevel,
+    onStreamMenuMicLevelChange,
+    streamMenuVolume,
+    onStreamMenuVolumeChange,
+    onStreamMenuToggleMicrophone,
+    onStreamMenuToggleFullscreen,
+    setCategoryIndex,
+    setSelectedSettingIndex,
+    setSettingsSubcategory,
+    setSelectedMediaIndex,
+    setMediaSubcategory,
+    setSelectedGameSubcategoryIndex,
+    setGameSubcategory,
+    setEditingBandwidth,
+    setEditingThemeChannel,
+    setEditingStreamVolume,
+    setEditingStreamMicLevel,
+    setOptionsEntries,
+    setOptionsOpen,
+    setOptionsFocusIndex,
+    setGamesHubFocusIndex,
+    setPs5Row,
+    setDetailRailIndex,
+    setGamesRootPlane,
+    setSpotlightIndex,
+    gamesHubReturnSnapshotRef,
+    setGamesHubOpen,
+    setEndSessionConfirm,
+    setLastRootGameIndex,
+    setLastRootSettingIndex,
+    setLastSystemMenuIndex,
+    setLastThemeRootIndex,
+    setLastRootMediaIndex,
+    setLibrarySortId,
+  } = ctx;
+
+  return useMemo(() => {
+    const eventContext: ControllerLibraryEventContext = {
+      isLoading,
+      TOP_CATEGORIES,
+      categorizedGames,
+      selectedIndex,
+      selectedGame,
+      selectedGameId,
+      selectedVariantId,
+      onPlayGame,
+      onSelectGameVariant,
+      onToggleFavoriteGame,
+      playUiSound,
+      throttledOnSelectGame,
+      topCategory,
+      selectedSettingIndex,
+      selectedMediaIndex,
+      selectedGameSubcategoryIndex,
+      displayItems,
+      mediaAssetItems,
+      mediaSubcategory,
+      gameSubcategory,
+      settings,
+      settingsSubcategory,
+      lastRootSettingIndex,
+      lastRootMediaIndex,
+      lastRootGameIndex,
+      lastSystemMenuIndex,
+      lastThemeRootIndex,
+      onSettingChange,
+      resolutionOptions,
+      fpsOptions,
+      codecOptions,
+      aspectRatioOptions,
+      currentStreamingGame,
+      onResumeGame,
+      onResumeCloudSession,
+      onCloseGame,
+      onExitControllerMode,
+      onExitApp,
+      editingBandwidth,
+      editingThemeChannel,
+      gamesShelfBrowseActive,
+      mediaShelfBrowseActive,
+      topLevelRowBehaviorActive,
+      topLevelShelfIndex,
+      canEnterDetailRow,
+      canEnterTopRow,
+      ps5Row,
+      detailRailIndex,
+      detailRailItems,
+      optionsOpen,
+      optionsFocusIndex,
+      optionsEntries,
+      gamesRootPlane,
+      spotlightIndex,
+      spotlightEntries,
+      gamesDualShelf,
+      favoriteGameIdSet,
+      microphoneDevices,
+      gamesHubOpen,
+      gamesHubFocusIndex,
+      gamesHubTiles,
+      inStreamMenu,
+      endSessionConfirm,
+      editingStreamVolume,
+      editingStreamMicLevel,
+      streamMenuMicLevel,
+      onStreamMenuMicLevelChange,
+      streamMenuVolume,
+      onStreamMenuVolumeChange,
+      onStreamMenuToggleMicrophone,
+      onStreamMenuToggleFullscreen,
+      setCategoryIndex,
+      setSelectedSettingIndex,
+      setSettingsSubcategory,
+      setSelectedMediaIndex,
+      setMediaSubcategory,
+      setSelectedGameSubcategoryIndex,
+      setGameSubcategory,
+      setEditingBandwidth,
+      setEditingThemeChannel,
+      setEditingStreamVolume,
+      setEditingStreamMicLevel,
+      setOptionsEntries,
+      setOptionsOpen,
+      setOptionsFocusIndex,
+      setGamesHubFocusIndex,
+      setPs5Row,
+      setDetailRailIndex,
+      setGamesRootPlane,
+      setSpotlightIndex,
+      gamesHubReturnSnapshotRef,
+      setGamesHubOpen,
+      setEndSessionConfirm,
+      setLastRootGameIndex,
+      setLastRootSettingIndex,
+      setLastSystemMenuIndex,
+      setLastThemeRootIndex,
+      setLastRootMediaIndex,
+      setLibrarySortId,
+    };
+    const { applyDirection, cycleTopCategory, onDirection, onShoulder } =
+      createNavigationHandlers(eventContext);
+    const {
+      onActivate,
+      onSecondaryActivate,
+      onTertiaryActivate,
+      onCancel,
+      onKeyboard,
+    } = createActionHandlers(eventContext, applyDirection, cycleTopCategory);
+    return {
+      onDirection: onDirection as (event: Event) => void,
+      onShoulder: onShoulder as (event: Event) => void,
+      onActivate,
+      onSecondaryActivate,
+      onTertiaryActivate,
+      onCancel,
+      onKeyboard,
+    };
+  }, [
+    isLoading,
+    TOP_CATEGORIES.length,
+    categorizedGames,
+    selectedIndex,
+    selectedGame,
+    selectedGameId,
+    selectedVariantId,
+    onPlayGame,
+    onSelectGameVariant,
+    onToggleFavoriteGame,
+    playUiSound,
+    throttledOnSelectGame,
+    topCategory,
+    selectedSettingIndex,
+    selectedMediaIndex,
+    selectedGameSubcategoryIndex,
+    displayItems,
+    mediaAssetItems,
+    mediaSubcategory,
+    gameSubcategory,
+    settings,
+    settingsSubcategory,
+    lastRootSettingIndex,
+    lastRootMediaIndex,
+    lastRootGameIndex,
+    lastSystemMenuIndex,
+    lastThemeRootIndex,
+    onSettingChange,
+    resolutionOptions,
+    fpsOptions,
+    codecOptions,
+    aspectRatioOptions,
+    currentStreamingGame,
+    onResumeGame,
+    onResumeCloudSession,
+    onCloseGame,
+    onExitControllerMode,
+    onExitApp,
+    editingBandwidth,
+    editingThemeChannel,
+    gamesShelfBrowseActive,
+    mediaShelfBrowseActive,
+    topLevelRowBehaviorActive,
+    topLevelShelfIndex,
+    canEnterDetailRow,
+    canEnterTopRow,
+    ps5Row,
+    detailRailIndex,
+    detailRailItems,
+    optionsOpen,
+    optionsFocusIndex,
+    optionsEntries.length,
+    gamesRootPlane,
+    spotlightIndex,
+    spotlightEntries,
+    gamesDualShelf,
+    favoriteGameIdSet,
+    microphoneDevices,
+    gamesHubOpen,
+    gamesHubFocusIndex,
+    gamesHubTiles,
+    inStreamMenu,
+    endSessionConfirm,
+    editingStreamVolume,
+    editingStreamMicLevel,
+    streamMenuMicLevel,
+    onStreamMenuMicLevelChange,
+    streamMenuVolume,
+    onStreamMenuVolumeChange,
+    onStreamMenuToggleMicrophone,
+    onStreamMenuToggleFullscreen,
+  ]);
+
+}
